+++ R/tm_t_pp_prior_medication.R
@@ -26,7 +26,7 @@ template_prior_medication <- function(dataname = "ANL",
   table_list <- add_expr(
     list(),
     substitute(expr = {
-      result <-
+      table_data <-
         dataname %>%
         dplyr::filter(atirel %in% c("PRIOR", "PRIOR_CONCOMITANT")) %>%
         dplyr::select(cmindc, cmdecod, cmstdy) %>%
@@ -34,20 +34,14 @@ template_prior_medication <- function(dataname = "ANL",
         dplyr::distinct() %>%
         `colnames<-`(teal.data::col_labels(dataname, fill = TRUE)[c(cmindc_char, cmdecod_char, cmstdy_char)])
 
-      table_listing <- result %>%
+      table <- table_data %>%
         dplyr::mutate( # Exception for columns of type difftime that is not supported by as_listing
           dplyr::across(
             dplyr::where(~ inherits(., what = "difftime")), ~ as.double(., units = "auto")
           )
         ) %>%
         rlistings::as_listing()
-
-      table <- DT::datatable(
-        table_listing,
-        options = list(
-          lengthMenu = list(list(-1, 5, 10, 25), list("All", "5", "10", "25"))
-        )
-      )
+      table
     }, env = list(
       dataname = as.name(dataname),
       atirel = as.name(atirel),
@@ -75,30 +69,6 @@ template_prior_medication <- function(dataname = "ANL",
 #' @inheritParams template_prior_medication
 #'
 #' @inherit module_arguments return
-#'
-#' @section Decorating Module:
-#'
-#' This module generates the following objects, which can be modified in place using decorators:
-#' - `table` (`datatables` - output of `DT::datatable()`)
-#'
-#' A Decorator is applied to the specific output using a named list of `teal_transform_module` objects.
-#' The name of this list corresponds to the name of the output to which the decorator is applied.
-#' See code snippet below:
-#'
-#' ```
-#' tm_t_pp_prior_medication(
-#'    ..., # arguments for module
-#'    decorators = list(
-#'      table = teal_transform_module(...) # applied only to `table` output
-#'    )
-#' )
-#' ```
-#'
-#' For additional details and examples of decorators, refer to the vignette
-#' `vignette("decorate-module-output", package = "teal.modules.clinical")`.
-#'
-#' To learn more please refer to the vignette
-#' `vignette("transform-module-output", package = "teal")` or the [`teal::teal_transform_module()`] documentation.
 #'
 #' @examplesShinylive
 #' library(teal.modules.clinical)
@@ -164,7 +134,16 @@ tm_t_pp_prior_medication <- function(label,
                                      pre_output = NULL,
                                      post_output = NULL,
                                      transformators = list(),
-                                     decorators = list()) {
+                                     decorators = lifecycle::deprecated()) {
+  if (lifecycle::is_present(decorators)) {
+    lifecycle::deprecate_warn(
+      when = "0.11.0",
+      what = "tm_t_pp_laboratory(decorators)",
+      details = "Decorators functionality was removed from this module. The `decorators` argument will be ignored."
+    )
+  }
+
+
   message("Initializing tm_t_pp_prior_medication")
   checkmate::assert_string(label)
   checkmate::assert_string(dataname)
@@ -176,7 +155,6 @@ tm_t_pp_prior_medication <- function(label,
   checkmate::assert_class(cmstdy, "choices_selected", null.ok = TRUE)
   checkmate::assert_class(pre_output, classes = "shiny.tag", null.ok = TRUE)
   checkmate::assert_class(post_output, classes = "shiny.tag", null.ok = TRUE)
-  assert_decorators(decorators, "table")
 
   args <- as.list(environment())
   data_extract_list <- list(
@@ -197,8 +175,7 @@ tm_t_pp_prior_medication <- function(label,
         dataname = dataname,
         parentname = parentname,
         label = label,
-        patient_col = patient_col,
-        decorators = decorators
+        patient_col = patient_col
       )
     ),
     transformators = transformators,
@@ -257,8 +234,7 @@ ui_t_prior_medication <- function(id, ...) {
         label = "Select CMSTDY variable:",
         data_extract_spec = ui_args$cmstdy,
         is_single_dataset = is_single_dataset_value
-      ),
-      ui_decorate_teal_data(ns("decorator"), decorators = select_decorators(ui_args$decorators, "table")),
+      )
     ),
     forms = tagList(
       teal.widgets::verbatim_popup_ui(ns("rcode"), button_label = "Show R code")
@@ -280,8 +256,7 @@ srv_t_prior_medication <- function(id,
                                    cmdecod,
                                    cmindc,
                                    cmstdy,
-                                   label,
-                                   decorators) {
+                                   label) {
   with_reporter <- !missing(reporter) && inherits(reporter, "Reporter")
   with_filter <- !missing(filter_panel_api) && inherits(filter_panel_api, "FilterPanelAPI")
   checkmate::assert_class(data, "reactive")
@@ -376,27 +351,24 @@ srv_t_prior_medication <- function(id,
         teal.code::eval_code(as.expression(unlist(my_calls)))
     })
 
-    # Decoration of table output.
-    decorated_table_q <- srv_decorate_teal_data(
-      id = "decorator",
-      data = all_q,
-      decorators = select_decorators(decorators, "table"),
-      expr = table
-    )
-
     # Outputs to render.
     table_r <- reactive({
-      q <- decorated_table_q()
-      list(
-        html = q[["table"]],
-        listing = q[["table_listing"]]
+      q <- all_q()
+
+      table_html <- DT::datatable(
+        data = q[["table"]],
+        options = list(
+          lengthMenu = list(list(-1, 5, 10, 25), list("All", "5", "10", "25"))
+        )
       )
+
+      list(html = table_html, report = q[["table"]])
     })
 
     output$prior_medication_table <- DT::renderDataTable(expr = table_r()$html)
 
     # Render R code.
-    source_code_r <- reactive(teal.code::get_code(req(decorated_table_q())))
+    source_code_r <- reactive(teal.code::get_code(req(all_q())))
     teal.widgets::verbatim_popup_srv(
       id = "rcode",
       verbatim_content = source_code_r,
@@ -412,10 +384,10 @@ srv_t_prior_medication <- function(id,
           filter_panel_api = filter_panel_api
         )
         card$append_text("Table", "header3")
-        if (nrow(table_r()$listing) == 0L) {
+        if (nrow(table_r()$report) == 0L) {
           card$append_text("No data available for table.")
         } else {
-          card$append_table(table_r()$listing)
+          card$append_table(table_r()$report)
         }
         if (!comment == "") {
           card$append_text("Comment", "header3")